// src/application/interfaces/the-movie-db.interface.ts
export interface GenreMovieApiResponse {
  genres: {
    id: number;
    name: string;
  }[];
}
export type DiscoverMovieApiResponse = {
  page: number;
  results: Movie[];
  total_pages: number;
  total_results: number;
};

export type Movie = {
  adult: boolean;
  backdrop_path: string;
  genre_ids: number[];
  id: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  release_date: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
};

export type MovieDetails = {
  id: number;
  title: string;
  genres: {
    id: number;
    name: string;
  }[];
};